Write README. Rename entities
