+++ raml-for-jaxrs/README.md
@@ -4,3 +4,87 @@ tags: ['Api Design', 'RAML', 'query params']
 license: 'cc-40-by'
 ---
 
+
+**NOT FOR PUBLISHING**
+
+
+## General
+
+The example project located in `example-proj` describes following features of `raml-to-jaxrs` (taken from https://github.com/mulesoft-labs/raml-for-jax-rs/blob/master/raml-to-jaxrs/README.md):
+
+* Generates interfaces from JSON Schema and RAML Types;
+* A response object wrapper is created for each resource action in order to guide the implementer in producing only results that are compatible with the RAML definition;
+* Custom annotations are generated for HTTP methods that are not part of the core JAX-RS specification;
+* Objects defined as RAML 1.0 objects are generated;
+* Objects are also generated based on JSON schemas to represent request/response entities;
+* This example uses Maven;
+
+
+## Example project structure
+
+*README.md*: Readme from original example project repo. May be useful to pull some info from it. Can be deleted.
+
+*pom.xml*: Required. I guess it's a java project setup file. Some info regarding its parts can be found in *README.md*.
+
+*src/main/java/server/ricks/RicksImpl.java*: Simple server to run generated JAX-RS project. Code can be demonstrated without it, but it's needed to demonstrate the project working. I guess users could create their own server but for demo purposes I think it may be good to include this one. Maybe its code should be listed after some title like "now let's create a simple server to test generated jax-rs server".
+
+*src/main/java/server/ricks/RicksImpl.java*: Required for server to run. Implements interfaces generated by the tool. This is logic for handling HTTP requests to the server.
+
+*src/main/resources/api.raml*: Required. API definition.
+
+*src/main/resources/includes/rick.json*: Required. JSON Scheme used in the API definition.
+
+*src/main/resources/includes/ricks-library.raml*: Required. RAML Library that defines type used in API definition.
+
+**Generated files (appear after generating jax-rs project):**
+
+*target/generated-sources/raml-to-jaxrs-maven-plugin/example/jsonschema/ResponseDelegate.java*: Looks like a delegate for Response. Is inherited in generated interfaces.
+
+*target/generated-sources/raml-to-jaxrs-maven-plugin/example/jsonschema/Rick.java*: Model generated from RAML Type. Defines setters/getters/etc.
+
+*target/generated-sources/raml-to-jaxrs-maven-plugin/example/jsonschema/Ricks.java*: API interface declaring all the HTTP methods, response classes.
+
+
+## To run the example:
+
+1. Install Maven
+2. In terminal run:
+
+```sh
+cd example-proj
+mvn install
+mvn raml:generate
+mvn exec:java
+```
+
+3. Now you have a server running at `localhost:9998`
+
+
+## Demo of the API/server
+
+http GET http://localhost:9998/ricks
+http POST http://localhost:9998/ricks slug=c137 name='Rick Sanchez' age:=70 inCouncil:=false
+http POST http://localhost:9998/ricks slug=riq name='Riq IV' age:=70 inCouncil:=true
+http POST http://localhost:9998/ricks slug=zar name='Zeta Alpha Rick' age:=70 inCouncil:=true
+http POST http://localhost:9998/ricks slug=prime name='Rick Prime' age:=60 inCouncil:=true
+http GET http://localhost:9998/ricks
+http GET http://localhost:9998/ricks/c137
+http GET http://localhost:9998/ricks/riq
+http GET http://localhost:9998/ricks/zar
+http GET http://localhost:9998/ricks/prime
+
+Same data in JSON if needed:
+
+{"slug": "c137", "name": "Rick Sanchez", "age": 70, "inCouncil": false}
+{"slug": "riq", "name": "Riq IV", "age": 70, "inCouncil": true}
+{"slug": "prime", "name": "Rick Prime", "age": 60, "inCouncil": true}
+{"slug": "zar", "name": "Zeta Alpha Rick", "age": 70, "inCouncil": true}
+
+
+## Things to note
+
+* Request validation does not work. I guess it's not supposed to work and code for it should not be generated by the tool;
+* Requesting unexisting urls does not return 404. I guess that's also supposed to be immplemented by the tool users;
+* As we discussed, `mvn install` step may be already run by Java IDEs;
+* Licences to `StartServer.java` and `RicksImpl.java` are added when running `mvn install` so they can be removed for article purposes;
+* Objects are stored in HashMap instead of real db;
