+++ raml-for-jaxrs/example-proj/README.md
@@ -1,9 +1,9 @@
-## Simple JSON Schema example
+## Simple example
 JSONSchema types are generated by [jsonschema2pojo](https://github.com/joelittlejohn/jsonschema2pojo)
 
 ```xml
 <configuration>
-    <ramlFile>${project.build.resources[0].directory}/example_json_schema.raml</ramlFile>
+    <ramlFile>${project.build.resources[0].directory}/api.raml</ramlFile>
     <resourcePackage>example.jsonschema</resourcePackage>
     <jsonMapper>jackson2</jsonMapper>
     <jsonMapperConfiguration>
@@ -34,16 +34,3 @@ These are the currently supported values for this configuration.
 * `DateType`
 
 You can find more information on this jsonschema2pojo documentation [page](http://joelittlejohn.github.io/jsonschema2pojo/site/0.4.14/generate-mojo.html).
-
-
-## To run example:
-
-1. Install Maven;
-2. In terminal run:
-
-```sh
-cd example_folder
-mvn install
-mvn raml:generate
-mvn exec:java
-```
