--- raml-for-jaxrs/example-proj/src/main/resources/includes/tracks-library.raml
@@ -1,12 +1,0 @@
-#%RAML 1.0 Library
-
-types:
-  Track:
-    type: object
-    properties:
-      id: string
-      title: string
-      albumId:
-        type: string
-        minLength: 36
-        maxLength: 36
